split the big project.js into smaller modules
