+++ dashboard/assets/styles/main.css
@@ -23,18 +23,11 @@ button.update-from-float {
     float: right;
 }
 
-.spinkit, .projects-spinkit, .graph-spinkit {
+.spinkit {
   position: relative;
   left: 45%;
   z-index: 9999;
-}
-
-.projects-spinkit {
   top: 200px;
-}
-
-.graph-spinkit {
-  top: 40px;
 }
 
 .status-header {
